+++ ARScrumWEBAPI/Controllers/SprintController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ARScrum.Services.Services.AppSprints;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 
 namespace ARScrumWEBAPI.Controllers
 {
@@ -13,10 +14,12 @@ namespace ARScrumWEBAPI.Controllers
     public class SprintController : ControllerBase
     {
         private readonly ISprintService _sprintService;
+        private readonly UserManager<AppUser> _userManager;
 
-        public SprintController(ISprintService sprintService)
+        public SprintController(ISprintService sprintService, UserManager<AppUser> userManager)
         {
             _sprintService = sprintService;
+            _userManager = userManager;
         }
 
         [HttpPost]
@@ -27,14 +30,21 @@ namespace ARScrumWEBAPI.Controllers
             {
                 return StatusCode(StatusCodes.Status403Forbidden, new AppResponse { Status = "Error", Message = "Sprint fields cannot be empty" });
             }
-            var result = await _sprintService.CreateSprintAsync(sprint);
-            if (result.IsSuccess && result.StatusCode == 200)
+            if (sprint.CreatedBy != "" || sprint.CreatedBy is not null)
             {
-                return Ok(result);
+                var user = await _userManager.FindByIdAsync(sprint.CreatedBy);
+                if (user is not null)
+                {
+                    var result = await _sprintService.CreateSprintAsync(sprint);
+                    if (result.IsSuccess && result.StatusCode == 200)
+                    {
+                        return Ok(result);
+                    }
+                    return StatusCode(StatusCodes.Status500InternalServerError, new AppResponse { Status = "Error", Message = result.Message });
+                }
+                return StatusCode(StatusCodes.Status403Forbidden, new AppResponse { Status = "Error", Message = "CreatedBy is not exist" });
             }
-
-            return StatusCode(StatusCodes.Status500InternalServerError, new AppResponse { Status = "Error", Message = result.Message });
-
+            return StatusCode(StatusCodes.Status400BadRequest, new AppResponse { Status = "Error", Message = "CreatedBy field cannot be empty" });
         }
 
         [HttpPost]
@@ -45,13 +55,21 @@ namespace ARScrumWEBAPI.Controllers
             {
                 return StatusCode(StatusCodes.Status403Forbidden, new AppResponse { Status = "Error", Message = "Sprint fields cannot be empty" });
             }
-            var result = await _sprintService.UpdateSprintAsync(sprint);
-            if (result.IsSuccess && result.StatusCode == 200)
+            if (sprint.UpdatedBy != "" || sprint.UpdatedBy is not null)
             {
-                return Ok(result);
+                var user = await _userManager.FindByIdAsync(sprint.UpdatedBy);
+                if (user is not null)
+                {
+                    var result = await _sprintService.UpdateSprintAsync(sprint);
+                    if (result.IsSuccess && result.StatusCode == 200)
+                    {
+                        return Ok(result);
+                    }
+                    return StatusCode(StatusCodes.Status500InternalServerError, new AppResponse { Status = "Error", Message = result.Message });
+                }
+                return StatusCode(StatusCodes.Status403Forbidden, new AppResponse { Status = "Error", Message = "UpdatedBy is not exist" });
             }
-
-            return StatusCode(StatusCodes.Status500InternalServerError, new AppResponse { Status = "Error", Message = result.Message });
+            return StatusCode(StatusCodes.Status400BadRequest, new AppResponse { Status = "Error", Message = "UpdatedBy field cannot be empty" });
 
         }
 
@@ -100,13 +118,21 @@ namespace ARScrumWEBAPI.Controllers
             {
                 return StatusCode(StatusCodes.Status403Forbidden, new AppResponse { Status = "Error", Message = "DeletedBy cannot be empty or null" });
             }
-
-            var result = await _sprintService.DeleteSprintAsync(sprintId, deletedBy);
-            if (result.IsSuccess && result.StatusCode == 200)
+            if (deletedBy != "" || deletedBy is not null)
             {
-                return Ok(result);
+                var user = await _userManager.FindByIdAsync(deletedBy);
+                if (user is not null)
+                {
+                    var result = await _sprintService.DeleteSprintAsync(sprintId, deletedBy);
+                    if (result.IsSuccess && result.StatusCode == 200)
+                    {
+                        return Ok(result);
+                    }
+                    return StatusCode(StatusCodes.Status500InternalServerError, new AppResponse { Status = "Error", Message = result.Message });
+                }
+                return StatusCode(StatusCodes.Status403Forbidden, new AppResponse { Status = "Error", Message = "DeletedBy is not exist" });
             }
-            return StatusCode(StatusCodes.Status500InternalServerError, new AppResponse { Status = "Error", Message = result.Message });
+            return StatusCode(StatusCodes.Status400BadRequest, new AppResponse { Status = "Error", Message = "DeletedBy field cannot be empty" });
         }
     }
 }
